Add Encryption.
Data in QRCodes are encrypted. The random encryption key is attached at top and bottom in paper.
This design provides some benefits:
1) when using a paper shredder, once the encryption key is destroyed, it's impossible to recover
the whole data. This reduces the working load to get a bundle of files be destroyed.
2) when data2paper is used for storing offline private keys(e.g. PGP), a user may distribute
the QR Codes to several people or location for redundancy, while still retaining his/her "main decryption key"
for being the decision maker in charge. This is in accordance with NeoAtlantis' plan of offline key management.